Conflicts:
	fish-web-monitor/src/main/webapp/app/locale/eway-locale-en.js

+++ fish-system/src/main/java/com/yihuacomputer/fish/person/service/db/OrganizationService.java
@@ -1,6 +1,7 @@
 package com.yihuacomputer.fish.person.service.db;
 
 import java.util.ArrayList;
+import java.util.Iterator;
 import java.util.List;
 
 import org.hibernate.Session;
@@ -43,6 +44,11 @@ public class OrganizationService extends DomainOrganizationService {
     private List<IOrganizationListener> orgListeners = new ArrayList<IOrganizationListener>();
 
     private final String SUB_ORG_HQL = "select o.id from Organization o where o.orgFlag like ? ";
+    
+    /**
+     * 如果为1则之前是根节点
+     */
+    private final static int ISLEAF=1;
 
     /**
      * 增加一条机构信息并初始化其编号和名称
@@ -71,10 +77,18 @@ public class OrganizationService extends DomainOrganizationService {
      * 增加一条机构信息
      */
     @Override
+    @Transactional
     public Organization add(IOrganization entity) {
         if (entity instanceof Organization) {
             Organization org = (Organization) entity;
             dao.save(entity);
+            IOrganization parent = org.getParent();
+            if(parent!=null){
+	            if(parent.isLeaf()){
+	            	parent.setLeaf(false);
+	            	dao.save(parent);
+	            }
+            }
             autoSetOrgFlag(org);
             dao.update(entity);
             return org;
@@ -159,7 +173,17 @@ public class OrganizationService extends DomainOrganizationService {
     @Override
     @CacheEvict(value = "orgs",key = "#guid")
     public void remove(String guid) {
+    	IOrganization org = this.get(guid);
+    	if(org==null){
+    		return;
+    	}
+    	IOrganization parent = org.getParent();
+    	
     	 dao.delete(Long.valueOf(guid), Organization.class);
+ 		if(!parent.listChildren().iterator().hasNext()){
+ 			parent.setLeaf(true);
+ 			dao.update(parent);
+ 		}
     }
 
     /**
@@ -170,13 +194,33 @@ public class OrganizationService extends DomainOrganizationService {
         Organization entity = interface2Entity(organization, true);
         String oldOrgFlag = entity.getOrgFlag();
         autoSetOrgFlag(entity);
-        //因为在hibernate中同一个session里面有了两个相同标识
-        Session session = sessionFactory.getCurrentSession();
+        //新的父节点
+        IOrganization newParent = entity.getParent();
+    	Session session = sessionFactory.getCurrentSession();
         session.clear();
         dao.update(entity);
+        //父节点的子节点包含新增的机构
+        List<IOrganization> childrenList = (ArrayList<IOrganization>)newParent.listChildren();
         String newOrgFlag = entity.getOrgFlag();
         // 级联修改子组织的orgFlag
         if (!oldOrgFlag.equalsIgnoreCase(newOrgFlag)) {
+        	//迁移后的机构数量为1，则要更改当前新父节点的leaf属性为false,非叶子节点
+            if(childrenList.size()==ISLEAF){
+                session.clear();
+            	newParent.setLeaf(false);
+            	dao.update(newParent);
+            }
+        	String oldFlags[] = oldOrgFlag.split("-");
+        	if(oldFlags.length>2){
+        		IOrganization oldParentOrg = this.get(oldFlags[oldFlags.length-2]);
+        		Iterator<IOrganization> orgIterable = oldParentOrg.listChildren().iterator();
+//        		orgIterable.next();
+        		if(!orgIterable.hasNext()){
+                    session.clear();
+        			oldParentOrg.setLeaf(true);
+        			dao.update(oldParentOrg);
+        		}
+        	}
             for (IOrganization child : entity.listChildren()) {
                 update(child);
             }
